+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Valve fills container only if aqueduct capacity has enough room to fill 
 - Configurable chance to consume water by Valve
 - Fixed Aqueduct doesn't have specified mineable tool
+- Configurable WellPulley player exhaustion on use
 
 ## [1.10.7]